@@ -16,7 +16,7 @@
         }
     ],
     "require": {
-        "php": ">=5.3.9",
+        "php": "^5.5.9|>=7.0.8",
         "symfony/polyfill-mbstring": "~1.0"
     },
     "require-dev": {
@@ -40,7 +40,7 @@
     "minimum-stability": "dev",
     "extra": {
         "branch-alias": {
-            "dev-master": "2.8-dev"
+            "dev-master": "3.3-dev"
         }
     }
 }
