@@ -1,16 +1,18 @@
 <?php
 
-/**
- * @file
- * Contains \Drupal\override_node_options\OverrideNodeOptionsNodeForm.
- */
-
 namespace Drupal\override_node_options\Form;
 
 use Drupal\node\NodeForm;
 use Drupal\Core\Form\FormStateInterface;
 
+/**
+ * Adds additional access checks for node override permissions.
+ */
 class OverrideNodeOptionsNodeForm extends NodeForm {
+
+  /**
+   * {@inheritdoc}
+   */
   protected function actions(array $form, FormStateInterface $form_state) {
     $element = parent::actions($form, $form_state);
     $node = $this->entity;
@@ -23,27 +25,29 @@ class OverrideNodeOptionsNodeForm extends NodeForm {
     if ($element['submit']['#access'] && $has_permissions) {
       // Add a "Publish" button.
       $element['publish'] = $element['submit'];
-      // If the "Publish" button is clicked, we want to update the status to "published".
+      // If the "Publish" button is clicked, we want to update the status to
+      // "published".
       $element['publish']['#published_status'] = TRUE;
       $element['publish']['#dropbutton'] = 'save';
       if ($node->isNew()) {
-        $element['publish']['#value'] = t('Save and publish');
+        $element['publish']['#value'] = $this->t('Save and publish');
       }
       else {
-        $element['publish']['#value'] = $node->isPublished() ? t('Save and keep published') : t('Save and publish');
+        $element['publish']['#value'] = $node->isPublished() ? $this->t('Save and keep published') : $this->t('Save and publish');
       }
       $element['publish']['#weight'] = 0;
 
       // Add a "Unpublish" button.
       $element['unpublish'] = $element['submit'];
-      // If the "Unpublish" button is clicked, we want to update the status to "unpublished".
+      // If the "Unpublish" button is clicked, we want to update the status to
+      // "unpublished".
       $element['unpublish']['#published_status'] = FALSE;
       $element['unpublish']['#dropbutton'] = 'save';
       if ($node->isNew()) {
-        $element['unpublish']['#value'] = t('Save as unpublished');
+        $element['unpublish']['#value'] = $this->t('Save as unpublished');
       }
       else {
-        $element['unpublish']['#value'] = !$node->isPublished() ? t('Save and keep unpublished') : t('Save and unpublish');
+        $element['unpublish']['#value'] = !$node->isPublished() ? $this->t('Save and keep unpublished') : $this->t('Save and unpublish');
       }
       $element['unpublish']['#weight'] = 10;
 
@@ -61,17 +65,18 @@ class OverrideNodeOptionsNodeForm extends NodeForm {
       $element['submit']['#access'] = FALSE;
     }
 
-    $element['preview'] = array(
+    $element['preview'] = [
       '#type' => 'submit',
       '#access' => $preview_mode != DRUPAL_DISABLED && ($node->access('create') || $node->access('update')),
-      '#value' => t('Preview'),
+      '#value' => $this->t('Preview'),
       '#weight' => 20,
-      '#submit' => array('::submitForm', '::preview'),
-    );
+      '#submit' => ['::submitForm', '::preview'],
+    ];
 
     $element['delete']['#access'] = $node->access('delete');
     $element['delete']['#weight'] = 100;
 
     return $element;
   }
+
 }
