@@ -35,14 +35,23 @@ class EntityVisibilityForm extends ConfigFormBase {
     // Get config for this form.
     $config = $this->config('entity_access_by_field.settings');
 
-    $form['disable_public_visibility'] = array(
+    $form['disable_public_visibility'] = [
       '#type' => 'checkbox',
       '#title' => $this->t('Disable public visibility'),
       '#description' => $this->t('Turning on this feature makes sure regular user can no longer post public content anywhere on the platform.'),
       '#default_value' => $config->get('disable_public_visibility'),
-    );
+    ];
 
-    // TODO: Change the autogenerated stub.
+    $form['default_visibility'] = [
+      '#type' => 'select',
+      '#title' => $this->t('Default visibility'),
+      '#default_value' => $config->get('default_visibility'),
+      '#options' => [
+        'public' => $this->t('Public - visible to everyone including people who are not a member'),
+        'community' => $this->t('Community - visible only to logged in members'),
+      ],
+    ];
+
     return parent::buildForm($form, $form_state);
   }
 
@@ -50,8 +59,10 @@ class EntityVisibilityForm extends ConfigFormBase {
    * {@inheritdoc}
    */
   public function submitForm(array &$form, FormStateInterface $form_state) {
-    $this->config('entity_access_by_field.settings')
-      ->set('disable_public_visibility', $form_state->getValue('disable_public_visibility'))
+    $form_state->cleanValues();
+    $this
+      ->config('entity_access_by_field.settings')
+      ->setData($form_state->getValues())
       ->save();
 
     parent::submitForm($form, $form_state);
