@@ -2,13 +2,14 @@
 
 namespace Drupal\webprofiler\State;
 
+use Drupal\Core\Cache\CacheCollector;
 use Drupal\Core\State\StateInterface;
 use Drupal\webprofiler\DataCollector\StateDataCollector;
 
 /**
  * Class StateWrapper.
  */
-class StateWrapper implements StateInterface {
+class StateWrapper extends CacheCollector implements StateInterface {
 
   /**
    * The system state.
@@ -90,6 +91,20 @@ class StateWrapper implements StateInterface {
    */
   public function resetCache() {
     $this->state->resetCache();
+  }
+
+  /**
+   * {@inheritdoc}
+   */
+  protected function resolveCacheMiss($key) {
+    return $this->state->resolveCacheMiss($key);
+  }
+
+  /**
+   * {@inheritdoc}
+   */
+  public function destruct() {
+    $this->updateCache();
   }
 
   /**
