@@ -54,10 +54,12 @@ class Quota extends MessagePurgeBase {
   /**
    * {@inheritdoc}
    */
-  public function fetch(MessageTemplateInterface $template, $limit) {
+  public function fetch(MessageTemplateInterface $template) {
     $query = $this->baseQuery($template);
     $result = $query
-      ->range($this->configuration['quota'], $limit)
+      // We need some kind of limit in order to get any results, but we really
+      // want all of them, so use an arbitrarily large number.
+      ->range($this->configuration['quota'], 1000000)
       ->execute();
     return $result;
   }
