@@ -8,7 +8,6 @@ use Drupal\search_api\IndexInterface;
 use Drupal\search_api\Processor\ProcessorPluginManager;
 use Drupal\search_api\SearchApiException;
 use Drupal\search_api\Tracker\TrackerPluginManager;
-use Symfony\Component\DependencyInjection\Exception\ServiceNotFoundException;
 
 /**
  * Provides methods for creating search plugins.
@@ -71,12 +70,12 @@ class PluginHelper implements PluginHelperInterface {
    *   Thrown if an unknown $type or $plugin_id is given.
    */
   protected function createIndexPlugin(IndexInterface $index, $type, $plugin_id, array $configuration = []) {
+    if (!isset($this->{$type . "PluginManager"})) {
+      throw new SearchApiException("Unknown plugin type '$type'");
+    }
     try {
       $configuration['#index'] = $index;
       return $this->{$type . "PluginManager"}->createInstance($plugin_id, $configuration);
-    }
-    catch (ServiceNotFoundException $e) {
-      throw new SearchApiException("Unknown plugin type '$type'");
     }
     catch (PluginException $e) {
       throw new SearchApiException("Unknown $type plugin with ID '$plugin_id'");
@@ -105,13 +104,11 @@ class PluginHelper implements PluginHelperInterface {
    *   Thrown if an unknown $type or plugin ID is given.
    */
   protected function createIndexPlugins(IndexInterface $index, $type, array $plugin_ids = NULL, array $configurations = []) {
+    if (!isset($this->{$type . "PluginManager"})) {
+      throw new SearchApiException("Unknown plugin type '$type'");
+    }
     if ($plugin_ids === NULL) {
-      try {
-        $plugin_ids = array_keys($this->{$type . "PluginManager"}->getDefinitions());
-      }
-      catch (ServiceNotFoundException $e) {
-        throw new SearchApiException("Unknown plugin type '$type'");
-      }
+      $plugin_ids = array_keys($this->{$type . "PluginManager"}->getDefinitions());
     }
 
     $plugins = [];
