@@ -11,7 +11,6 @@ use Drupal\search_api\Entity\Server;
 use Drupal\search_api\Item\Field;
 use Drupal\search_api\Processor\ProcessorInterface;
 use Drupal\search_api_test\PluginTestTrait;
-use Drupal\Tests\taxonomy\Functional\TaxonomyTestTrait;
 
 /**
  * Tests the admin UI for processors.
@@ -24,7 +23,6 @@ class ProcessorIntegrationTest extends SearchApiBrowserTestBase {
 
   use EntityReferenceTestTrait;
   use PluginTestTrait;
-  use TaxonomyTestTrait;
 
   /**
    * {@inheritdoc}
@@ -115,6 +113,13 @@ class ProcessorIntegrationTest extends SearchApiBrowserTestBase {
 
     $this->checkContentAccessIntegration();
     $enabled[] = 'content_access';
+    sort($enabled);
+    $actual_processors = array_keys($this->loadIndex()->getProcessors());
+    sort($actual_processors);
+    $this->assertEquals($enabled, $actual_processors);
+
+    $this->checkEntityBundleBoostIntegration();
+    $enabled[] = 'type_boost';
     sort($enabled);
     $actual_processors = array_keys($this->loadIndex()->getProcessors());
     sort($actual_processors);
@@ -371,6 +376,29 @@ class ProcessorIntegrationTest extends SearchApiBrowserTestBase {
       unset($field_settings['label'], $field_settings['dependencies']);
       $this->assertEquals($settings, $field_settings, "Field $field_id has the correct settings.");
     }
+  }
+
+  /**
+   * Tests the UI for the "Type-specific boosting" processor.
+   */
+  public function checkEntityBundleBoostIntegration() {
+    $configuration = [
+      'boosts' => [
+        'entity:node' => [
+          'datasource_boost' => '3.0',
+          'bundle_boosts' => [
+            'article' => '5.0',
+          ],
+        ],
+        'entity:user' => [
+          'datasource_boost' => '1.0',
+        ],
+      ],
+    ];
+    $form_values = $configuration;
+    $form_values['boosts']['entity:node']['bundle_boosts']['page'] = '';
+
+    $this->editSettingsForm($configuration, 'type_boost', $form_values);
   }
 
   /**
