@@ -17,17 +17,17 @@ use Symfony\Component\DependencyInjection\ContainerInterface;
  * )
  */
 class Entity implements MentionsPluginInterface {
-  private $token_service;
-  private $entity_manager;
-  private $entityQuery_service;
+  private $tokenService;
+  private $entityManager;
+  private $entityQueryService;
 
   /**
    * Entity constructor.
    */
   public function __construct(Token $token, EntityManager $entity_manager, QueryFactory $entity_query) {
-    $this->token_service = $token;
-    $this->entity_manager = $entity_manager;
-    $this->entityQuery_service = $entity_query;
+    $this->tokenService = $token;
+    $this->entityManager = $entity_manager;
+    $this->entityQueryService = $entity_query;
   }
 
   /**
@@ -48,11 +48,11 @@ class Entity implements MentionsPluginInterface {
    * {@inheritdoc}
    */
   public function outputCallback($mention, $settings) {
-    $entity = $this->entity_manager->getStorage($mention['target']['entity_type'])
+    $entity = $this->entityManager->getStorage($mention['target']['entity_type'])
       ->load($mention['target']['entity_id']);
-    $output['value'] = $this->token_service->replace($settings['value'], array($mention['target']['entity_type'] => $entity));
+    $output['value'] = $this->tokenService->replace($settings['value'], array($mention['target']['entity_type'] => $entity));
     if ($settings['renderlink']) {
-      $output['link'] = $this->token_service->replace($settings['rendertextbox'], array($mention['target']['entity_type'] => $entity));
+      $output['link'] = $this->tokenService->replace($settings['rendertextbox'], array($mention['target']['entity_type'] => $entity));
     }
     return $output;
   }
@@ -63,7 +63,7 @@ class Entity implements MentionsPluginInterface {
   public function targetCallback($value, $settings) {
     $entity_type = $settings['entity_type'];
     $input_value = $settings['value'];
-    $query = $this->entityQuery_service->get($entity_type);
+    $query = $this->entityQueryService->get($entity_type);
     $result = $query->condition($input_value, $value)->execute();
 
     if (!empty($result)) {
