@@ -3,13 +3,17 @@
 namespace Drupal\social_profile\Plugin\Block;
 
 use Drupal\Core\Block\BlockBase;
+use Drupal\Core\Cache\Cache;
 
 /**
  * Provides a 'ProfileHeroBlock' block.
  *
  * @Block(
- *  id = "profile_hero_block",
- *  admin_label = @Translation("Profile hero block"),
+ *   id = "profile_hero_block",
+ *   admin_label = @Translation("Profile hero block"),
+ *   context = {
+ *     "user" = @ContextDefinition("entity:user")
+ *   }
  * )
  */
 class ProfileHeroBlock extends BlockBase {
@@ -19,33 +23,45 @@ class ProfileHeroBlock extends BlockBase {
    */
   public function build() {
     $build = [];
+    $account = $this->getContextValue('user');
 
-    $account = \Drupal::routeMatch()->getParameter('user');
-    if (!is_object($account) && !is_null($account)) {
-      $account = \Drupal::service('entity_type.manager')
-        ->getStorage('user')
-        ->load($account);
+    $storage = \Drupal::entityTypeManager()->getStorage('profile');
+    $profile = $storage->loadByUser($account, 'profile');
+
+    if ($profile) {
+      $build['content'] = \Drupal::entityTypeManager()
+        ->getViewBuilder('profile')
+        ->view($profile, 'hero');
+      $build['content']['#cache']['tags'] = $this->getCacheTags();
+      $build['content']['#cache']['contexts'] = $this->getCacheContexts();
     }
-
-    if (!empty($account)) {
-      $storage = \Drupal::entityTypeManager()->getStorage('profile');
-      if (!empty($storage)) {
-        $user_profile = $storage->loadByUser($account, 'profile');
-        if ($user_profile) {
-          $content = \Drupal::entityTypeManager()
-            ->getViewBuilder('profile')
-            ->view($user_profile, 'hero');
-          $build['content'] = $content;
-        }
-      }
-    }
-
-    // @TODO cache.
-    $build['#cache'] = array(
-      'max-age' => 0,
-    );
 
     return $build;
+  }
+
+  /**
+   * {@inheritdoc}
+   */
+  public function getCacheTags() {
+    $account = $this->getContextValue('user');
+    $storage = \Drupal::entityTypeManager()->getStorage('profile');
+    $profile = $storage->loadByUser($account, 'profile');
+    $tags = [
+      'user:' . $account->id(),
+    ];
+
+    if ($profile) {
+      $tags[] = 'profile:' . $profile->id();
+    }
+
+    return Cache::mergeTags(parent::getCacheTags(), $tags);
+  }
+
+  /**
+   * {@inheritdoc}
+   */
+  public function getCacheContexts() {
+    return Cache::mergeContexts(parent::getCacheContexts(), ['user.permissions']);
   }
 
 }
