@@ -39,7 +39,7 @@ class ExportUser extends ActionBase implements ContainerFactoryPluginInterface {
    *
    * @var bool
    */
-  protected $apply_all;
+  protected $applyAll;
 
   /**
    * The query.
@@ -86,7 +86,7 @@ class ExportUser extends ActionBase implements ContainerFactoryPluginInterface {
    * {@inheritdoc}
    */
   public function executeMultiple(array $entities) {
-    if ($this->apply_all) {
+    if ($this->applyAll) {
       $this->tempStoreFactory->get('user_operations_export')->set($this->currentUser->id(), [
         'apply_all' => TRUE,
         'query' => $this->query,
@@ -117,11 +117,11 @@ class ExportUser extends ActionBase implements ContainerFactoryPluginInterface {
   /**
    * Set the apply_all property.
    *
-   * @param bool $apply_all
+   * @param bool $applyAll
    *   Boolean that determines if this applies for all or not.
    */
-  public function setApplyAll($apply_all) {
-    $this->apply_all = $apply_all;
+  public function setApplyAll($applyAll) {
+    $this->applyAll = $applyAll;
   }
 
   /**
